+++ test/support/router.ex
@@ -5,19 +5,25 @@ defmodule TestAppWeb.PostController do
   def call(conn, _opts), do: conn
 
   def create(conn, %{"post" => post_params}) do
-    changeset =
+    _changeset =
       %Post{}
       |> Ecto.Changeset.cast(post_params, [:title, :body, :published])
 
     conn
+    |> Plug.Conn.put_status(:created)
+    |> Phoenix.Controller.render(:show, post: %Post{})
   end
 
-  def update(conn, %{"id" => id, "post" => post_params}) do
-    changeset =
+  def update(conn, %{"id" => _id, "post" => post_params}) do
+    _changeset =
       %Post{}
       |> Ecto.Changeset.cast(post_params, [:title, :body])
 
-    conn
+    Phoenix.Controller.render(conn, :show, post: %Post{})
+  end
+
+  def delete(conn, %{"id" => _id}) do
+    Plug.Conn.send_resp(conn, :no_content, "")
   end
 end
 
@@ -46,7 +52,7 @@ defmodule TestAppWeb.Router do
   scope "/api", TestAppWeb do
     pipe_through :api
 
-    resources "/posts", PostController, only: [:index, :show, :create]
+    resources "/posts", PostController, only: [:index, :show, :create, :update, :delete]
     resources "/users", UserController, only: [:index, :show]
 
     resources "/posts/:post_id/comments", CommentController, only: [:index, :show]